+++ imageinf/inference/__init__.py
@@ -1,0 +1,10 @@
+from .config import DEFAULT_MODEL_NAME
+from .processor import run_model_on_tapis_images
+from .registry import MODEL_REGISTRY, MODEL_METADATA
+
+__all__ = [
+    "DEFAULT_MODEL_NAME",
+    "run_model_on_tapis_images",
+    "MODEL_REGISTRY",
+    "MODEL_METADATA",
+]
