+++ imageinf/inference/config.py
@@ -1,0 +1,1 @@
+DEFAULT_MODEL_NAME = "google/vit-base-patch16-224"
