+++ imageinf/utils/config.py
@@ -1,0 +1,1 @@
+CACHE_DIR = "cache_images"  # TODO add periodic cleanup
